RDKE-419: getMtlscert temporary change to support community devices
Reason for Change: getMtlscert is not implemented in a generic way
and it was returning SUCCESS always which is causing runtime issues
on MTLS not supported devices. Community devices does not support
LIBRDKCONFIG_BUILD as well. Use that as a temporary solution until
RDKE-419 gets proper solution.
Test Procesure: Verify the firmware upgrade feature.